+++ tests/Integration/TestCase.php
@@ -29,6 +29,7 @@ abstract class TestCase extends WPTestUtilsTestCase {
 	 * Makes sure the foundational stuff is sorted so tests work.
 	 */
 	public function set_up() {
+		parent::set_up();
 
 		// We need to trick the plugin into thinking it's run by WP-CLI.
 		if ( ! defined( 'WP_CLI' ) ) {